# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "defaultDescription",
    "version" : "0.1",
    "title" : "defaultTitle"
  },
  "paths" : {
    "/student" : {
      "get" : {
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "$ref" : "#/definitions/Model5"
            }
          }
        }
      },
      "post" : {
        "consumes" : [ "application/json" ],
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model0"
          },
          "x-examples" : {
            "application/json" : "{\n    \"first_name\":\"ahmad\",\n    \"last_name\":\"adnan\",\n    \"college\":\"tafila\",\n    \"phone_number\":\"0773330824\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/student/{param0}" : {
      "get" : {
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "param0",
          "in" : "path",
          "description" : "Example values form path are: 'amman' and 'meaan'",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "$ref" : "#/definitions/Model6"
            }
          }
        }
      }
    },
    "/student/{param1}" : {
      "delete" : {
        "consumes" : [ "application/json" ],
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model2"
          },
          "x-examples" : {
            "application/json" : "{\n    \"first_name\":\"ahmad\",\n    \"last_name\":\"adnan\",\n    \"college\":\"tafila\",\n    \"phone_number\":\"0773330824\"\n}"
          }
        }, {
          "name" : "param1",
          "in" : "path",
          "description" : "Example values form path are: '22' and '1'",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "type" : "string"
            }
          },
          "400" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/student/{param2}" : {
      "put" : {
        "consumes" : [ "application/json" ],
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model1"
          },
          "x-examples" : {
            "application/json" : "{\n    \"first_name\":\"osama\",\n    \"last_name\":\"adnan\",\n    \"college\":\"amman\",\n    \"phone_number\":\"0773330824\"\n}"
          }
        }, {
          "name" : "param2",
          "in" : "path",
          "description" : "Example values form path are: '22' and '2'",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "type" : "string"
            }
          },
          "400" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Model0" : {
      "properties" : {
        "first_name" : {
          "type" : "string"
        },
        "last_name" : {
          "type" : "string"
        },
        "college" : {
          "type" : "string"
        },
        "phone_number" : {
          "type" : "string"
        }
      }
    },
    "Model1" : {
      "properties" : {
        "first_name" : {
          "type" : "string"
        },
        "last_name" : {
          "type" : "string"
        },
        "college" : {
          "type" : "string"
        },
        "phone_number" : {
          "type" : "string"
        }
      }
    },
    "Model2" : {
      "properties" : {
        "first_name" : {
          "type" : "string"
        },
        "last_name" : {
          "type" : "string"
        },
        "college" : {
          "type" : "string"
        },
        "phone_number" : {
          "type" : "string"
        }
      }
    },
    "Array" : {
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "first_name" : {
          "type" : "string"
        },
        "last_name" : {
          "type" : "string"
        },
        "college" : {
          "type" : "string"
        },
        "phone_number" : {
          "type" : "string"
        }
      }
    },
    "Model5" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/Array"
      }
    },
    "Model6_Array" : {
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "first_name" : {
          "type" : "string"
        },
        "last_name" : {
          "type" : "string"
        },
        "college" : {
          "type" : "string"
        },
        "phone_number" : {
          "type" : "string"
        }
      }
    },
    "Model6" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/Model6_Array"
      }
    }
  }
}